+++ src/main/kotlin/com/github/vitallium/rubylsp/RubyLspServerDescriptor.kt
@@ -96,7 +96,9 @@ internal class RubyLspServerDescriptor(
                 gemScriptArgsBuilder.add("--experimental")
             }
 
-            return RubyGemExecutionContext.tryCreate(RubySdkUtil.getFileSdk(project, file), null, GEM_SCRIPT_NAME)
+            val moduleForRunningLsp = if (lspSettings.useBundler) module else null
+
+            return RubyGemExecutionContext.tryCreate(RubySdkUtil.getFileSdk(project, file), moduleForRunningLsp, GEM_SCRIPT_NAME)
                 ?.withWorkingDir(gemfile.parent)
                 ?.withGemScriptName(GEM_SCRIPT_NAME)
                 ?.withArguments(gemScriptArgsBuilder)